traffic_MVP_project/requirements.txt
ultralytics>=8.3.0
opencv-python>=4.8.0
pyyaml>=6.0
numpy>=1.26
pandas>=2.0
streamlit>=1.30
